Initial refactor of the project
